refactor(memory): call the fork state-carry directly; retire the persistence-hooks seam
onConversationForked was the last member of the memoryPersistenceHooks
seam object. Extract its body into a standalone plugin function,
forkConversationMemory (new file fork-conversation-memory.ts), call it
directly from the fork path in conversation-crud.ts, and delete
persistence-hooks.ts entirely. The "hooks" naming was a misnomer — this
is a direct plugin call, not a hooks-pipeline dispatch.

It stays a direct import (not a first-class hooks event like the other
lifecycle notifications) because it runs synchronously inside the fork DB
transaction, threading the live transaction handle so the child's memory
state commits atomically with the fork — a shape the async hooks pipeline
cannot express. The fork mechanism is expected to change before any hook
is exposed for it, so this keeps the coupling honest and direct in the
meantime rather than dressing it up as a lifecycle event.

Layering: the persistence-layering guard allowlist entry for
conversation-crud.ts swaps "persistence-hooks" for
"fork-conversation-memory" (the "indexer" entry from the addMessage
inlining is unchanged). Also corrected a stale comment in pkb-schedule.ts
that referenced the deleted seam — hasPkbBufferContent is consumed
directly by jobs-worker.ts and filing-jobs.ts, never through the seam.
