+++ assistant/src/plugins/defaults/memory/pkb-schedule.ts
@@ -2,9 +2,8 @@
  * The PKB buffer's "is there filable work?" check, as a leaf module (fs +
  * workspace path only).
  *
- * The jobs-worker's maintenance scheduler consults it through the
- * `MemoryPersistenceHooks` seam (see `persistence-hooks.ts`) when deciding
- * whether to enqueue a `pkb_filing` job, and the filing handler
+ * The jobs-worker's maintenance scheduler (`jobs-worker.ts`) consults it when
+ * deciding whether to enqueue a `pkb_filing` job, and the filing handler
  * (`filing-jobs.ts`) re-checks it at run time.
  */
